chore: rename jwtgenerator to jujuauth
